Initial import of the proof-of-concept.
The spring boot application uses React as template engine to render
the index page. It parses index.html to execute exactly the same
javascript code that was generated by the create-react-app production
build.

build-client.sh on the client runs the production build (it runs
npm run build) and then moves and copies the created files to the
correct locations in the server directory.

The spring boot app uses ScriptEngine to execute the JavaScript code.
I decided to not change it to the Polyglot interface of Graalvm for
now, because I want to make more progress on the ReactJS part.