Routing: Controller gets requestURI from ServletRequest
Playing around with routing. The first thing I need to implement
routing on the server side is the complete request URI - or at least
the part after the context.

This was step 1 - the easy part: Added the servlet API as a dependency
to the build. Then I was able to inject the HttpServletRequest to
the controller and get the URI from it.