Improve first-load: Defer all scripts
To improve the first-load performance, I added code to defer all scripts
in Controller.kt. Now, when the script to render the application is
loaded, the other scripts might not be loaded yet; So, I had to add
code to index.html to periodically check if the function is already
there.