Loads list of items from the server when running with npm.
The React App now loads a list of items from the server - BUT this
uses the fetch API for now, so it only works when running the app
with npm!

So, the server has a new controller for "/api/list" where it serves
the hard-coded list ['a', 'b'] (in ApiController.kt). The react app
fetches this list and displays it as a <ul/>.

In package.json, I set up proxying the spring boot server on
localhost, so that there are no CORS issues when calling this API.