Server returns dynamic list, can add item.
The app now has and input field to add new items. An "Add" button
next to that field will post the new item to the server, and after
completion, load the list of items again.

On the server, the ApiController.kt maintains a mutable list of
items instead of the static list ['a', 'b'] from the last commit.
It adds new items that are posted to /api/add and returns this list
when getting /api/list.