List uses data from server on full reload.
When re-loading the list with CTRL+F5, the useEffect(...) of the
list app does not fetch the list JSON again - it uses the same
data that was used on the server.

To enable that, onServer creates a script tag that is rendered
into the list component **only on the server**. This script tag
initializes the value window.serverData['app.list'] with the
data that was used when rendering the list of the server.

In the list component, the useEffect(...) first checks whether
serverData('app.list') already contains the data. Only if there
is no data (i.e. when there was a dynamic navigation), the list
component's effect fetches the JSON from the server.